+++ README.md
@@ -21,7 +21,7 @@
 
 1. Open up `android/app/src/main/java/[...]/MainApplication.java`
 
-- Add `import nabidreams.media.AudioPackage;` to the imports at the top of the file
+- Add `import com.nabidreams.lib.audio.AudioPackage;` to the imports at the top of the file
 - Add `new AudioPackage()` to the list returned by the `getPackages()` method
 
 2. Append the following lines to `android/settings.gradle`:
