+++ android/src/main/AndroidManifest.xml
@@ -1,4 +1,4 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-          package="nabidreams.media">
+          package="com.nabidreams.lib.audio">
 
 </manifest>
