+++ android/src/main/java/com/nabidreams/lib/audio/AudioPackage.kt
@@ -1,4 +1,4 @@
-package nabidreams.media
+package com.nabidreams.lib.audio
 
 import java.util.Arrays
