+++ example/android/app/BUCK
@@ -35,12 +35,12 @@ android_library(
 
 android_build_config(
     name = "build_config",
-    package = "nabidreams.media.audioexample",
+    package = "com.nabidreams.lib.audio.example",
 )
 
 android_resource(
     name = "res",
-    package = "nabidreams.media.audioexample",
+    package = "com.nabidreams.lib.audio.example",
     res = "src/main/res",
 )
