+++ example/android/app/google-services.json
@@ -8,9 +8,9 @@
   "client": [
     {
       "client_info": {
-        "mobilesdk_app_id": "1:240002467684:android:af508e5ee22f8e5f775c71",
+        "mobilesdk_app_id": "1:240002467684:android:6926a9057b85556d775c71",
         "android_client_info": {
-          "package_name": "nabidreams.media.audioexample"
+          "package_name": "com.nabidreams.lib.audio.example"
         }
       },
       "oauth_client": [
@@ -30,6 +30,13 @@
             {
               "client_id": "240002467684-i95b7ib0f4ob0dji2l87ge1nukchh0lq.apps.googleusercontent.com",
               "client_type": 3
+            },
+            {
+              "client_id": "240002467684-o6n6a9dv2r3i78mf6uc3n494knu9s2aq.apps.googleusercontent.com",
+              "client_type": 2,
+              "ios_info": {
+                "bundle_id": "com.nabidreams.lib.audio.example"
+              }
             }
           ]
         }
