+++ example/android/app/src/main/java/com/nabidreams/lib/audio/example/MainActivity.java
@@ -1,4 +1,4 @@
-package nabidreams.media.audioexample;
+package com.nabidreams.lib.audio.example;
 
 import com.facebook.react.ReactActivity;
