+++ example/android/fastlane/Appfile
@@ -1,2 +1,2 @@
 json_key_file "" # Path to the json secret file - Follow https://github.com/fastlane/supply#setup to get one
-package_name "nabidreams.media.audioexample" # e.g. com.krausefx.app
+package_name "com.nabidreams.lib.audio.example" # e.g. com.krausefx.app
